requirements.txt
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.50
botocore==1.40.50
certifi==2025.10.5
cffi==2.0.0
cryptography==46.0.2
fastapi==0.118.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
google-cloud-compute==1.20.0
google-auth==2.35.0
GitPython==3.1.43
psycopg2-binary==2.9.9
